requirements.txt
ipykernel==6.26.0
ipython==8.17.2
openai==1.2.3
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.5.2
python-dotenv==1.0.0
urllib3==2.0.7

